% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChartData_ExpectationRanges}
\alias{ChartData_ExpectationRanges}
\title{Chart data template expectation ranges}
\format{
An object of class \code{data.frame} with 36 rows and 13 columns.
}
\usage{
ChartData_ExpectationRanges
}
\description{
Chart data template expectation ranges
}
\keyword{datasets}
