% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChartData_FinalEnergy}
\alias{ChartData_FinalEnergy}
\title{Chart data template final energy}
\format{
An object of class \code{data.frame} with 24 rows and 16 columns.
}
\usage{
ChartData_FinalEnergy
}
\description{
Chart data template final energy
}
\keyword{datasets}
