% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChartData_HeatNeed}
\alias{ChartData_HeatNeed}
\title{Chart data template heat need}
\format{
An object of class \code{data.frame} with 12 rows and 16 columns.
}
\usage{
ChartData_HeatNeed
}
\description{
Chart data template heat need
}
\keyword{datasets}
