% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChartSettings_ExpectationRanges}
\alias{ChartSettings_ExpectationRanges}
\title{Chart settings expectation ranges}
\format{
An object of class \code{data.frame} with 4 rows and 28 columns.
}
\usage{
ChartSettings_ExpectationRanges
}
\description{
Chart settings expectation ranges
}
\keyword{datasets}
