% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChartSettings_FinalEnergy}
\alias{ChartSettings_FinalEnergy}
\title{Chart settings final energy}
\format{
An object of class \code{data.frame} with 4 rows and 28 columns.
}
\usage{
ChartSettings_FinalEnergy
}
\description{
Chart settings final energy
}
\keyword{datasets}
