% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChartSettings_HeatNeed}
\alias{ChartSettings_HeatNeed}
\title{Chart settings heat need}
\format{
An object of class \code{data.frame} with 5 rows and 26 columns.
}
\usage{
ChartSettings_HeatNeed
}
\description{
Chart settings heat need
}
\keyword{datasets}
