% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulaChartsFunctions.R
\name{LoadChartSettings}
\alias{LoadChartSettings}
\title{Load charts settings}
\usage{
LoadChartSettings(
  mySourceType = "RDA",
  mySheetName,
  myFileName = "ChartSettings"
)
}
\arguments{
\item{mySourceType}{The type of data source; possible entries: "RDA" (default), "Excel"}

\item{mySheetName}{The name of the Excel sheet that includes the data table to be loaded}

\item{myFileName}{The name of the file that includes the data table to be loaded}
}
\value{
a dataframe of the type "ChartData" or "ChartSettings"
}
\description{
The function loads the settings of the chart.
}
\examples{
# Load the "ChartData" template and the chart settings from an Excel file located 
# in "Input/Template/Excel/"  


# 1 Chart displaying the energy balance of the building fabric and the heat need 

ChartData_HeatNeed <- 
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_HeatNeed",
    mySheetName =     "DF_ChartData"
  )
     
ChartSettings_HeatNeed <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_HeatNeed",
    mySheetName =     "DF_Settings"
  )

head (ChartData_HeatNeed)
head (ChartSettings_HeatNeed)


# 2 Chart displaying the final energy demand 

ChartData_FinalEnergy <- 
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_FinalEnergy",
    mySheetName =     "DF_ChartData"
  )

ChartSettings_FinalEnergy <- 
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_FinalEnergy",
    mySheetName =     "DF_Settings"
  )

head (ChartData_FinalEnergy)
head (ChartSettings_FinalEnergy)


# 3 Chart displaying expectation ranges of the heat need and the final energy demand

ChartData_ExpectationRanges <- 
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_Uncertainties",
    mySheetName =     "DF_ChartData"
  )

ChartSettings_ExpectationRanges <- 
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_Uncertainties",
    mySheetName =     "DF_Settings"
  )
 
 head (ChartData_ExpectationRanges)
 head (ChartSettings_ExpectationRanges)
 
 
}
