% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulaChartsFunctions.R
\name{LoadExcelChartParameters}
\alias{LoadExcelChartParameters}
\title{Load charts settings and templates from Excel}
\usage{
LoadExcelChartParameters(myDataFrameType = "ChartSettings", myChartType)
}
\arguments{
\item{myDataFrameType}{The dataframe type: "ChartSettings" (default) or "ChartDataTemplate"}

\item{myCharType}{The chart type: "HeatNeed", "FinalEnergy" or "ExpectationRanges"}
}
\value{
a dataframe of the type "ChartData" or "ChartSettings"
}
\description{
The function loads the settings of the chart.
}
\examples{
# Load the chart data template and the chart settings  
# from an Excel file located at "Input/Excel/"  

# (1) Chart displaying the energy balance of the building fabric and the heat need 

ChartSettings_HeatNeed <-
  LoadExcelChartParameters (
    myDataFrameType = "ChartSettings",
    myChartType     = "HeatNeed"
  )

ChartData_HeatNeed <- 
  LoadExcelChartParameters (
    myDataFrameType = "ChartDataTemplate",
    myChartType     = "HeatNeed"
  )
     
 head (ChartSettings_HeatNeed)
 head (ChartData_HeatNeed)


# (2) Chart displaying the final energy demand 

ChartSettings_FinalEnergy <-
  LoadExcelChartParameters (
    myDataFrameType = "ChartSettings",
    myChartType     = "FinalEnergy"
  )

ChartData_FinalEnergy <- 
  LoadExcelChartParameters (
    myDataFrameType = "ChartDataTemplate",
    myChartType     = "FinalEnergy"
  )
     
 head (ChartSettings_FinalEnergy)
 head (ChartData_FinalEnergy)


# (3) Chart displaying expectation ranges of the heat need and the final energy demand

ChartSettings_ExpectationRanges <-
  LoadExcelChartParameters (
    myDataFrameType = "ChartSettings",
    myChartType     = "ExpectationRanges"
  )

ChartData_ExpectationRanges <- 
  LoadExcelChartParameters (
    myDataFrameType = "ChartDataTemplate",
    myChartType     = "ExpectationRanges"
  )
     
 head (ChartSettings_ExpectationRanges)
 head (ChartData_ExpectationRanges)
 
 # Export the RDA files 
 # Files can later be copied to the subedirectory "data" to be included in the R package
 
 save (ChartSettings_HeatNeed,          file = "ChartSettings_HeatNeed.rda")
 save (ChartSettings_FinalEnergy,       file = "ChartSettings_FinalEnergy.rda")
 save (ChartSettings_ExpectationRanges, file = "ChartSettings_ExpectationRanges.rda")
 
 save (ChartData_HeatNeed,              file = "ChartData_HeatNeed.rda")
 save (ChartData_FinalEnergy,           file = "ChartData_FinalEnergy.rda")
 save (ChartData_ExpectationRanges,     file = "ChartData_ExpectationRanges.rda")
 
 
}
