% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulaChartsFunctions.R
\name{ShowBarChart}
\alias{ShowBarChart}
\title{Show a bar chart of the energy performance calculation}
\usage{
ShowBarChart(
  myChartSettings,
  myChartData,
  DF_EnergyData,
  Index_Dataset = 1,
  Code_Language = "ENG",
  Set_MaxY_Auto = TRUE,
  Do_FlipChart = FALSE,
  stackStrategy = "samesign"
)
}
\arguments{
\item{myChartSettings}{A dataframe including the chart settings}

\item{myChartData}{A dataframe including the data structure, labels, chart colours, sources}

\item{DF_EnergyData}{A dataframe including the energy data to be displayed}

\item{Index_Dataset}{An integer indicating which dataset (row of DF_EnergyData) to be displayed}

\item{Code_Language}{A character string to indicate the language of the chart labels;
English: "ENG", German = "GER", other language = "XXX"
The labels are entered in the respective columns in "myChartData"}

\item{Set_MaxY_Auto}{A boolean indicating if automatic scaling is to be applied
or if predefined MaxY values from "myChartSettings" should be used}

\item{Do_FlipChart}{A boolean indicating if the bar chart should be flipped
from vertical columns to horizonal bars}

\item{stackStrategy}{A character string indicating the handling of negative values;
the options are:
'samesign': only stack values if the value to be stacked has the same sign as the currently cumulated stacked value.
'all': stack all values, irrespective of the signs of the current or cumulative stacked value.
'positive': only stack positive values.
'negative': only stack negative values.
see: https://echarts.apache.org/en/option.html#series-bar.stack}
}
\value{
A bar chart. In order to show the chart in the browser call: options (viewer = NULL)
}
\description{
The function uses the output from the package MobasyModel to create a bar chart
to the dataframe template for the heat need data.
}
\examples{

## Only call this if you want to show the chart in the browser 
options (viewer = NULL) 


## Load the energy data
load ("Input/Data/Example/myOutputTables.rda") 
# This is a dataframe generated by the functions MobasyCalc () and EnergyProfileCalc () 
# of the R package MobasyModel
DF_EnergyData <-
  myOutputTables$DF_Display_Energy

## 1 Heat need
## Chart displaying the energy balance of the building fabric and the heat need
## Example building: "DE.MOBASY.WBG.0008.05"

ChartData_HeatNeed <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_HeatNeed",
    mySheetName =     "DF_ChartData"
  )

ChartSettings_HeatNeed <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_HeatNeed",
    mySheetName =     "DF_Settings"
  )

ShowBarChart  (
  myChartSettings = ChartSettings_HeatNeed,
  myChartData     = ChartData_HeatNeed,
  DF_EnergyData   = DF_EnergyData,
  Index_Dataset   = which (DF_EnergyData$ID_Dataset == "DE.MOBASY.WBG.0008.05"),
  Code_Language   = "GER",
  #Code_Language  = "ENG",
  Set_MaxY_Auto   = TRUE,
  Do_FlipChart    = FALSE,
  stackStrategy   = 'samesign'
)


## 2 Final energy
## Chart displaying the final energy demand
## Example building: "DE.MOBASY.NH.0033.05"

ChartData_FinalEnergy <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_FinalEnergy",
    mySheetName =     "DF_ChartData"
  )

ChartSettings_FinalEnergy <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_FinalEnergy",
    mySheetName =     "DF_Settings"
  )

ShowBarChart  (
  myChartSettings = ChartSettings_FinalEnergy,
  myChartData     = ChartData_FinalEnergy,
  DF_EnergyData   = DF_EnergyData,
  Index_Dataset   = which (DF_EnergyData$ID_Dataset == "DE.MOBASY.NH.0033.05"),
  Code_Language   = "GER",
  #Code_Language  = "ENG",
  Set_MaxY_Auto   = TRUE,
  Do_FlipChart    = FALSE,
  stackStrategy   = 'samesign'
)


## 3 Expectation ranges
## Chart displaying expectation ranges of the heat need and the final energy demand
## Example building: "DE.MOBASY.NH.0033.05"

ChartData_ExpectationRanges <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_Uncertainties",
    mySheetName =     "DF_ChartData"
  )

ChartSettings_ExpectationRanges <-
  LoadChartSettings (
    mySourceType =    "Excel",
    myFileName =      "ChartSettings_Uncertainties",
    mySheetName =     "DF_Settings"
  )

ShowBarChart  (
  myChartSettings = ChartSettings_ExpectationRanges,
  myChartData     = ChartData_ExpectationRanges,
  DF_EnergyData   = DF_EnergyData,
  Index_Dataset   = which (DF_EnergyData$ID_Dataset == "DE.MOBASY.NH.0033.05"),
  Code_Language   = "GER",
  #Code_Language   = "ENG",
  Set_MaxY_Auto   = TRUE,
  Do_FlipChart    = TRUE,
  stackStrategy   = 'all'
)


}
