% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulaChartsFunctions.R
\name{ShowBarChart}
\alias{ShowBarChart}
\title{Show a bar chart of the energy performance calculation}
\usage{
ShowBarChart(
  myChartSettings,
  myChartData,
  DF_EnergyData,
  Index_Dataset = 1,
  Code_Language = "ENG",
  Set_MaxY_Auto = TRUE,
  Do_FlipChart = FALSE,
  stackStrategy = "samesign",
  ScalingFactor_FontSize = 1
)
}
\arguments{
\item{myChartSettings}{A dataframe including the chart settings}

\item{myChartData}{A dataframe including the data structure, labels, chart colours, sources}

\item{DF_EnergyData}{A dataframe including the energy data to be displayed}

\item{Index_Dataset}{An integer indicating which dataset (row of DF_EnergyData) to be displayed}

\item{Code_Language}{A character string to indicate the language of the chart labels;
English: "ENG", German = "GER", other language = "XXX"
The labels are entered in the respective columns in "myChartData"}

\item{Set_MaxY_Auto}{A boolean indicating if automatic scaling is to be applied
or if predefined MaxY values from "myChartSettings" should be used}

\item{Do_FlipChart}{A boolean indicating if the bar chart should be flipped
from vertical columns to horizonal bars}

\item{stackStrategy}{A character string indicating the handling of negative values;
the options are:
'samesign': only stack values if the value to be stacked has the same sign as the currently cumulated stacked value.
'all': stack all values, irrespective of the signs of the current or cumulative stacked value.
'positive': only stack positive values.
'negative': only stack negative values.
see: https://echarts.apache.org/en/option.html#series-bar.stack}

\item{ScalingFactor_FontSize}{A real value for scaling the font size (default value = 1.0)}
}
\value{
A bar chart. In order to show the chart in the browser call: options (viewer = NULL)
}
\description{
The function uses the output from the package MobasyModel to create a bar chart
to the dataframe template for the heat need data.
}
\examples{

## Only call this if you want to show the chart in the browser 
options (viewer = NULL) 


## Load the energy data
load ("Input/RDA/MobasyOutputData/Example/myOutputTables.rda") 
# This is a dataframe generated by the functions MobasyCalc () and EnergyProfileCalc () 
# of the R package MobasyModel
DF_EnergyData <-
  myOutputTables$DF_Display_Energy

## (1) Heat need
## Chart displaying the energy balance of the building fabric and the heat need
## Example building: "DE.MOBASY.WBG.0008.05"


# Load the chart settings  
ChartSettings_HeatNeed <- TabulaCharts::ChartSettings_HeatNeed

# Load the template for the chart data   
ChartData_HeatNeed     <- TabulaCharts::ChartData_HeatNeed


## Alternatively load both from the respective Excel file

ChartSettings_HeatNeed <-
  LoadExcelChartParameters (
    myDataFrameType = "ChartSettings",
    myChartType     = "HeatNeed"
  )

ChartData_HeatNeed <- 
  LoadExcelChartParameters (
    myDataFrameType = "ChartDataTemplate",
    myChartType     = "HeatNeed"
  )
     
## Create and show the chart

ShowBarChart  (
  myChartSettings = ChartSettings_HeatNeed [2, ],
  myChartData     = ChartData_HeatNeed,
  DF_EnergyData   = DF_EnergyData,
  Index_Dataset   = which (DF_EnergyData$ID_Dataset == "DE.MOBASY.WBG.0008.05"),
  Code_Language   = "GER",
  #Code_Language  = "ENG",
  Set_MaxY_Auto   = TRUE,
  Do_FlipChart    = FALSE,
  stackStrategy   = 'samesign',
  ScalingFactor_FontSize = 1.0
)


## (2) Final energy
## Chart displaying the final energy demand
## Example building: "DE.MOBASY.NH.0033.05"

# Load the chart settings  
ChartSettings_FinalEnergy <- TabulaCharts::ChartSettings_FinalEnergy

# Load the template for the chart data   
ChartData_FinalEnergy     <- TabulaCharts::ChartData_FinalEnergy


## Alternatively load both from the respective Excel file

ChartSettings_FinalEnergy <-
  LoadExcelChartParameters (
    myDataFrameType = "ChartSettings",
    myChartType     = "FinalEnergy"
  )
  
ChartData_FinalEnergy <- 
  LoadExcelChartParameters (
    myDataFrameType = "ChartDataTemplate",
    myChartType     = "FinalEnergy"
  )
     
ShowBarChart  (
  myChartSettings = ChartSettings_FinalEnergy [2, ],
  myChartData     = ChartData_FinalEnergy,
  DF_EnergyData   = DF_EnergyData,
  Index_Dataset   = which (DF_EnergyData$ID_Dataset == "DE.MOBASY.NH.0033.05"),
  Code_Language   = "GER",
  #Code_Language  = "ENG",
  Set_MaxY_Auto   = TRUE,
  Do_FlipChart    = FALSE,
  stackStrategy   = 'samesign',
  ScalingFactor_FontSize = 1.0
)


## 3 Expectation ranges
## Chart displaying expectation ranges of the heat need and the final energy demand
## Example building: "DE.MOBASY.NH.0033.05"


# Load the chart settings  
ChartSettings_ExpectationRanges <- TabulaCharts::ChartSettings_ExpectationRanges

# Load the template for the chart data   
ChartData_ExpectationRanges     <- TabulaCharts::ChartData_ExpectationRanges


## Alternatively load both from the respective Excel file

ChartSettings_ExpectationRanges <-
  LoadExcelChartParameters (
    myDataFrameType = "ChartSettings",
    myChartType     = "ExpectationRanges"
  )

ChartData_ExpectationRanges <- 
  LoadExcelChartParameters (
    myDataFrameType = "ChartDataTemplate",
    myChartType     = "ExpectationRanges"
  )
     
ShowBarChart  (
  myChartSettings = ChartSettings_ExpectationRanges [2, ],
  myChartData     = ChartData_ExpectationRanges,
  DF_EnergyData   = DF_EnergyData,
  Index_Dataset   = which (DF_EnergyData$ID_Dataset == "DE.MOBASY.NH.0033.05"),
  Code_Language   = "GER",
  #Code_Language   = "ENG",
  Set_MaxY_Auto   = TRUE,
  Do_FlipChart    = TRUE,
  stackStrategy   = 'all',
  ScalingFactor_FontSize = 1.0
)


}
